- upgrade to spring 2.0.5 and hibernate 3.2.3.ga, ehcache 1.2.3 - remove beanRefFactory from context. This avoids scanning beans twice at startup. - rename localDataSource.xml -> dataSource.xml. (generated file) - upgrade to compass 1.1.1, cglib 2.1_3 - fix stray dependencies to old hibernate and spring (I think) - rewrite the probe viewer. The new version is reachable from the genecontroller. - deprecate GeneDao.getCoexpressedGeneMap (nobody was using it) - remove foo.bar.id=la.de.id in many HQL queries (this is not valid and you can just do foo.bar=la.de) - fix other hql queries that have incorrect syntax in new hibernate version - fix sloooow query to get taxon for array design - fix UserDetails serialization (?) - add timing to startuplistener - bioMaterialAnnotator - start of annotation tool. - remove 'findById' methods from services (please use 'load'  and reserve find* for actual finders)

+++ gemma-mda/src/main/java/ubic/gemma/model/expression/bioAssay/BioAssayDaoImpl.java
@@ -84,7 +84,7 @@ public class BioAssayDaoImpl extends ubic.gemma.model.expression.bioAssay.BioAss
         try {
             org.hibernate.Query queryObject = super.getSession( false ).createQuery( query );
 
-            return ( Integer ) queryObject.iterate().next();
+            return (( Long ) queryObject.iterate().next()).intValue();
         } catch ( org.hibernate.HibernateException ex ) {
             throw super.convertHibernateAccessException( ex );
         }
